% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bamregion2GRanges.R
\name{bamregion2GRanges}
\alias{bamregion2GRanges}
\title{Overrides StrandPhaseR's function for loading sequence data}
\usage{
bamregion2GRanges(
  bamfile,
  bamindex,
  region = NULL,
  pairedEndReads = FALSE,
  min.mapq = 10,
  filterAltAlign = TRUE,
  all_alignments = galignmentslist_global_for_invertyper
)
}
\value{
a GRanges object
}
\description{
This replaces the StrandPhaseR function by sneakily preventing it from reading from the BAM file, but instead giving it processed GAlignments results
A list of GAlignments objects, named by BAM file name, will need to be available to this function even though it's not passed from StrandPhaseR
}
