% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{boundaries}
\alias{boundaries}
\title{Finds peaks in deltaW values}
\usage{
boundaries(df)
}
\arguments{
\item{df}{A dataframe produced by GenomicRanges::mergeByOverlaps. Column 1 contains Granges objects (inversions), column 2 contains Granges objects that overlap the
objects in column 1 (reads) and the remaining columns are \code{mcols}, of which one must be \code{$deltaW}.}
}
\value{
A dataframe containing genomic coordinates of the two peaks in deltaW values
}
\description{
Intermediate function that processes the input and output of the twopeaks function.
}
\details{
Extracts a vector of deltaW values from the dataframe of GRanges objects, feeds it to twopeaks, and then extracts the genomic coordinates of the peaks that are returned.
}
