% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/breakpointr_for_invertyper.R
\name{breakpointr_for_invertyper}
\alias{breakpointr_for_invertyper}
\title{A stripped down parallel implementation of the core BreakpointR function, "runBreakpointr".
Lots of this code is borrowed/altered from the BreakpointR package by David Porubsky, Ashley Sanders, and Aaron Taudt.
It also doesn't do such nice validation of arguments
Needs import foreach, doParallel, and maybe utils from read.table
Probably need to ad BreakpointR:: to all the BPR functions
We get a separate PDF for each file... maybe that could be improved?
qPDF could be used for that
plotspath should be a directory, not a filename
it uses single end GRanges with 2 frags paired.
For plotting, should also downsample beforehand to reduce processing time. Plus----sort the GRANGES! otherwise the chromosomes are all mixed up.}
\usage{
breakpointr_for_invertyper(
  grangeslist = NULL,
  paired_reads = FALSE,
  plotspath = NULL,
  numCPU = 4,
  windowsize = 1e+06,
  binMethod = "size",
  minReads = 50,
  background = 0.2,
  maskRegions = NULL,
  chromosomes = NULL
)
}
\arguments{
\item{chromosomes}{}
}
\value{
\link{...}
}
\description{
A stripped down parallel implementation of the core BreakpointR function, "runBreakpointr".
Lots of this code is borrowed/altered from the BreakpointR package by David Porubsky, Ashley Sanders, and Aaron Taudt.
It also doesn't do such nice validation of arguments
Needs import foreach, doParallel, and maybe utils from read.table
Probably need to ad BreakpointR:: to all the BPR functions
We get a separate PDF for each file... maybe that could be improved?
qPDF could be used for that
plotspath should be a directory, not a filename
it uses single end GRanges with 2 frags paired.
For plotting, should also downsample beforehand to reduce processing time. Plus----sort the GRANGES! otherwise the chromosomes are all mixed up.
}
