% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/choose_priors.R
\name{choose_priors}
\alias{choose_priors}
\title{Helps find reasonable priors for inversion genotyping}
\usage{
choose_priors(regions_to_genotype, expected_number_inversions = 100)
}
\arguments{
\item{regions_to_genotype}{A GRanges object of putative inversions. If you have a BED file, see import_bed()}

\item{expected_number_inversions}{An integer. If you had perfect data and genotyped an average individual using your regions_to_genotype, how many
unique inversions do you think there would be? For example, most humans have something like 100-200 paracentric inversions. If you think your inversion list
probably covers a lot of them, then maybe 100 is a good estimate. On the other hand, if your inversion list is a long shot and probably only a few of them are real,
you could consider a smaller number}
}
\value{
A list with the recommended prior and haploid_prior
}
\description{
With InvertypeR it is possible to genotype very long, messy lists of putative inversions from the literature, or from genomic features like inverted repeats
Sometimes it is difficult to choose a prior in these cases. What we do here is guess the number of inversions we think are present in the list of putative
inversions, and use that plus self-intersections of the inversion list to come up with a reasonable prior.
}
\details{
This is not intended for cases where the inversion list does not have self-intersections----for example, if you are just genotyping a handful of well-characterized
inversions. Then you should create a prior based on known allele frequencies or something.

This function could give ridiculous suggestions for priors if there are regions_to_genotype, for example. Use your judgement.
}
