% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_regions.R
\name{count_regions}
\alias{count_regions}
\title{Counts reads in genomic regions}
\usage{
count_regions(WW_reads, WC_reads, regions)
}
\arguments{
\item{regions}{A GRanges object for putative inversions}

\item{WW_bam}{A GRanges object for the WW composite file}

\item{WC_bam}{A GRanges object for the WC composite file}
}
\value{
A dataframe with read counts by file
}
\description{
For each composite file, this function tallies up the forward and reverse reads in each putative inversion (region).
}
\details{
One of the slowest steps of the genotyping process when composite files are large.
}
