% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_composite_files.R
\name{create_composite_files}
\alias{create_composite_files}
\title{Creates composite files for inversion genotyping}
\usage{
create_composite_files(
  inputfolder = "./",
  type = c("wc", "ww"),
  numCPU = 24,
  vcf = NULL,
  paired_reads = TRUE,
  blacklist = NULL,
  outputfolder = "./",
  save_composite_files = FALSE,
  chromosomes = NULL
)
}
\arguments{
\item{chromosomes}{}
}
\value{
\link{...}
}
\description{
Creates composite files for inversion genotyping
}
