% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discover_possible_inversions.R
\name{discover_possible_inversions}
\alias{discover_possible_inversions}
\title{Discovering putative inversions as strand-switches with BreakpointR}
\usage{
discover_possible_inversions(
  composite_files = list(),
  windowsize = c(40, 120, 260),
  minReads = c(15, 50, 50),
  paired_reads = FALSE,
  numCPU = 4,
  chromosomes = NULL,
  hard_mask = NULL,
  background = 0.2,
  type = c("wc", "ww")
)
}
\arguments{
\item{composite_files}{A list of 1 or 2 composite files, named WC and WW (or just WW for haploids)}

\item{windowsize}{A vector of integers. Each integer will be used the set the number of reads in a bin for a BreakpointR run.}

\item{minReads}{A vector of integers parallel to windowsize, giving the minimum number of reads in a bin for it to be considered/genotyped}

\item{paired_reads}{Boolean. Paired-end reads?}

\item{numCPU}{Integer. How many threads to use?}

\item{chromosomes}{A character vector of chromosome names to examine}

\item{hard_mask}{A GRanges object containing intervals with suspected poor-quality Strand-seq data.}

\item{background}{A parameter to be passed on to BreakpointR, setting the maximum amount of background allowable for ww or cc calls.}

\item{type}{'wc', 'ww', or both. Describes the input composite files (which are either Watson-Watson or Watson-Crick, or both).}
}
\value{
A GRanges object of potential inversions for genoptyping with invertyper()
}
\description{
Discovering putative inversions as strand-switches with BreakpointR
}
