% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{drop_seq_qual}
\alias{drop_seq_qual}
\title{Removes the 'seq' and 'qual' mcols from a GAlignments or GAlignmentsPairs object}
\usage{
drop_seq_qual(galignments, paired_reads = TRUE)
}
\arguments{
\item{galignments}{A GAlignments object}

\item{paired_reads}{Boolean}
}
\value{
A GAlignments object without the offending mcols
}
\description{
I'm concerned that these fields are burdensome to store, so I want to get rid of them in the standard implementation
}
