% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{extract_reads_by_region_and_state}
\alias{extract_reads_by_region_and_state}
\title{Extracts reads from a GAlignments object that match a certain strand state within a given region}
\usage{
extract_reads_by_region_and_state(
  galignmentslist = NULL,
  regions = NULL,
  paired_reads = TRUE,
  states = "wc",
  filename = NULL,
  flip_reads = FALSE
)
}
\arguments{
\item{galignmentslist}{A list of GAlignments objects}

\item{regions}{A GRanges object with a 'states' mcol and a 'filename' mcol matching the GAlignments objects}

\item{paired_reads}{Boolean}

\item{states}{The desired strand state(s)}

\item{filename}{The filename for the GAlignments object in the list from which reads should be extracted}

\item{flip_reads}{Whether to reorient reads for composite file creation (e.g., reorient cc regions for a ww composite file)}
}
\value{
A GAlignments object
}
\description{
Extracts reads from a GAlignments object that match a certain strand state within a given region
}
