% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{find_regions_with_strand_state}
\alias{find_regions_with_strand_state}
\title{Using BreakpointR output, this identifies regions of a particular strand state
For example, Watson-Crick regions for phasing}
\usage{
find_regions_with_strand_state(
  list_of_breakpoint_objects = list(),
  states = c("wc"),
  region_size = 1e+05
)
}
\arguments{
\item{list_of_breakpoint_objects}{A list of Breakpoint objects from BreakpointR}

\item{states}{The desired Strand-state ('ww', 'wc', or 'cc')}

\item{region_size}{The minimum size of regions to select}
}
\value{
A GRanges object containing genomic intervals of the desired strand state(s)
}
\description{
Using BreakpointR output, this identifies regions of a particular strand state
For example, Watson-Crick regions for phasing
}
