% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/galignment_to_granges.R
\name{galignment_to_granges}
\alias{galignment_to_granges}
\title{Converting reads from GAlignments to GRanges formats}
\usage{
galignment_to_granges(
  galignments = NULL,
  purpose = NULL,
  paired_reads = TRUE,
  region = NULL,
  pair2frgm = FALSE
)
}
\arguments{
\item{galignments}{A GenomicAlignments object}

\item{purpose}{Either 'StrandPhaseR' or 'BreakpointR'. Different purposes result in different formats.}

\item{paired_reads}{Boolean. Are the reads paired?}

\item{region}{A GRanges object, to which the GAlignments object can be subset.}

\item{pair2frgm}{Boolean. Should paired reads be merged into a single fragment, for BreakpointR?}
}
\value{
A GRanges object
}
\description{
The resulting GRanges objects are suitable for processing with other Strand-seq R packages like StrandPhaseR or BreakpointR
}
