% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_reads.R
\name{import}
\alias{import}
\title{A utility function to import BED files}
\usage{
import(path)
}
\arguments{
\item{path}{Path to a BED file}
}
\value{
A GRanges object
}
\description{
Reads the first 3 columns of a BED file into a GRanges object.
}
