% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_bam.R
\name{import_bam}
\alias{import_bam}
\title{Reads a BAM file into a GAlignments object, subsetting to regions/chromosomes of interest}
\usage{
import_bam(
  bam = NULL,
  region = NULL,
  paired_reads = TRUE,
  blacklist = NULL,
  chromosomes = NULL
)
}
\arguments{
\item{bam}{Path to a BAM file. Alternatively, a GRanges or GAlignments object}

\item{region}{A GRanges object of regions from which reads should be taken. Default NULL.}

\item{paired_reads}{Boolean. Are the reads paired?}

\item{blacklist}{A GRanges object of regions from which reads should NOT be taken. Default NULL}

\item{chromosomes}{A character vector of chromosome names. Reads from other chromosomes will not be loaded. Default NULL (all chromosomes).}
}
\value{
A GAlignments object of reads
}
\description{
It can also take a GAlignments or GRanges object, in which case it will just apply the necessary region subsetting instead of trying to load it fresh somehow.
}
