% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/invertyper_pipeline.R
\name{invertyper_pipeline}
\alias{invertyper_pipeline}
\title{The BED files assume 1-based coordinates.}
\usage{
invertyper_pipeline(
  regions_to_genotype = NULL,
  prior = c(0.333, 0.333, 0.333),
  haploid_prior = c(0.5, 0.5),
  adjust_method = c("raw", "merge", "deltas", "minimal", "low", "all"),
  inputfolder = "./",
  outputfolder = "./",
  haploid_chromosomes = NULL,
  vcf = NULL,
  paired_reads = TRUE,
  confidence = 0.95,
  blacklist = NULL,
  chromosomes = NULL,
  numCPU = 24,
  save_composite_files = FALSE,
  write_browser_files = FALSE,
  discover_breakpointr_inversions = TRUE,
  breakpointr_prior = c(0.9, 0.05, 0.05),
  breakpointr_haploid_prior = c(0.9, 0.1),
  windowsize = c(40, 120, 360),
  minReads = c(15, 50, 50),
  background = 0.2,
  output_file = "inversions.txt"
)
}
\arguments{
\item{output_file}{}
}
\value{
\link{...}
}
\description{
The BED files assume 1-based coordinates.
}
