% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadMatrices.R
\name{loadMatrices}
\alias{loadMatrices}
\title{Copied from StrandPhaseR so that it accesses InvertypeR's version of bamregion2granges
The function is unchanged (except for @inheritParams), and it's just here for namespace finagling}
\usage{
loadMatrices(
  inputfolder = NULL,
  positions = NULL,
  WCregions = NULL,
  pairedEndReads = FALSE,
  min.mapq = 10,
  min.baseq = 30
)
}
\description{
This funcion will read in partial single cell haplotypes into
two parallel matrices separately for Watson and Crick reads
}
\author{
David Porubsky
}
