% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadMatrices_for_invertyper.R
\name{loadMatrices_for_invertyper}
\alias{loadMatrices_for_invertyper}
\title{Copied from StrandPhaseR so that it accesses InvertypeR's version of bamregion2granges
The function is unchanged (except for @inheritParams, the bamregions function, and some added package:: tags), and it's just here for namespace finagling}
\usage{
loadMatrices_for_invertyper(
  input_folder = NULL,
  positions = NULL,
  WCregions = NULL,
  pairedEndReads = FALSE,
  min.mapq = 10,
  min.baseq = 30
)
}
\arguments{
\item{input_folder}{Path to the bam files to process}

\item{positions}{Filename with listed position of SNVs for given chromosome (format: chrName SNVpos).}

\item{WCregions}{Filename of all WC region for a given chromosome (format: chrName:Start:End:FileName).}

\item{pairedEndReads}{Set to \code{TRUE} if you have paired-end reads in your file.}

\item{min.mapq}{Minimum mapping quality when importing from BAM files.}

\item{min.baseq}{Minimum base quality to consider a base for phasing.}
}
\description{
This funcion will read in partial single cell haplotypes into
two parallel matrices separately for Watson and Crick reads
}
\author{
David Porubsky
}
