% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{minimal}
\alias{minimal}
\title{Finds the intersection of a set of overlapping intervals}
\usage{
minimal(interval)
}
\arguments{
\item{interval}{A GRanges object}
}
\value{
A GRanges object
}
\description{
Used in an experimental inversion adjustment routine.
}
\details{
For a GRanges object containing multiple clusters of overlapping intervals, this function acts on each cluster. For each one, it finds the largest interval that overlaps every interval in the
cluster. When this is not possible, it returns a placeholder interval of length zero.
}
