% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_reads.R
\name{pair2frgm}
\alias{pair2frgm}
\title{Mandatory preprocessing of reads for BreakpointR (Porubsky et al. 2020)}
\usage{
pair2frgm(data.raw)
}
\arguments{
\item{data.raw}{A GAlignmentPairs object}
}
\value{
A GenomicRanges object
}
\description{
Code from BreakpointR, https://doi.org/10.1093/bioinformatics/btz681
}
\details{
To use the deltaWCalculator from BreakpointR with paired-end reads, they must first be merged into one fragement. It's not supplied as a function so we have to reproduce a piece of the code.
}
