% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phaseChromosome_for_invertyper.R
\name{phaseChromosome_for_invertyper}
\alias{phaseChromosome_for_invertyper}
\title{A stripped-down version of the StrandPhaseR function phaseChromosome}
\usage{
phaseChromosome_for_invertyper(
  inputfolder,
  outputfolder = "./StrandPhaseR_analysis",
  positions = NULL,
  WCregions = NULL,
  chromosome = NULL,
  pairedEndReads = TRUE,
  min.mapq = 10,
  min.baseq = 30,
  num.iterations = 2,
  translateBases = TRUE,
  fillMissAllele = NULL,
  splitPhasedReads = FALSE,
  compareSingleCells = FALSE,
  exportVCF = NULL
)
}
\description{
This is a simplified function that removes some of the functionality (and runtime) of David Porubsky's original StrandPhaseR function
}
