% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_reads.R
\name{process_reads}
\alias{process_reads}
\title{Read processing for inversion adjustment}
\usage{
process_reads(WW_reads, WC_reads, paired_reads = TRUE)
}
\arguments{
\item{WW_reads}{A GRanges object from the WW composite file.}

\item{WC_reads}{A GRanges object from the WC composite file.}

\item{paired_reads}{Boolean: are the reads paired-end? Default TRUE.}
}
\value{
A list of four GRanges objects: the input objects WW_reads and WC_reads and the two output objects (for WW and WC composite files) that have been annotated with deltaW values.
}
\description{
Input reads are sent through the deltaWCalculator from BreakpointR, which associates them with deltaW values approximating the local change in read direction
}
\details{
For each read, we sum multiple deltaW values calculated for binsizes 5, 10, 10, 20, 40, and 80. Yes, 10 is there twice: it means we're prioritizing changes in read direction from
the smaller scale.
}
