% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_bam.R
\name{read_bam}
\alias{read_bam}
\title{For composite BAM writing, maybe something related to fwrite could be created? Fastest way to write... write to SAM format and then compress? Or use rtracklayer's function
Going to expect sorted, indexed BAM files.
Maybe for SNV calling, https://bioconductor.org/packages/release/bioc/manuals/VariantTools/man/VariantTools.pdf should be used?
However, that looks clumsy: it won't take GAlingments format. So it's best to ask for an external VCF, and just suggest in the documentation that this can easily be done with bbtools (give command)
If a GenomicAlignments or GRanges object is passed instead of the path to a BAM file, this will just try excluding reads that aren't part of the provided regions. This is helpful because
That way InvertypeR can be given either a GAlignments composite file (to throw away useless reads) or a BAM composite file (old-school InvertypeR)}
\usage{
read_bam(
  bam = NULL,
  region = NULL,
  paired_reads = TRUE,
  blacklist = NULL,
  chromosomes = NULL
)
}
\arguments{
\item{chromosomes}{}
}
\value{
\link{...}
}
\description{
For composite BAM writing, maybe something related to fwrite could be created? Fastest way to write... write to SAM format and then compress? Or use rtracklayer's function
Going to expect sorted, indexed BAM files.
Maybe for SNV calling, https://bioconductor.org/packages/release/bioc/manuals/VariantTools/man/VariantTools.pdf should be used?
However, that looks clumsy: it won't take GAlingments format. So it's best to ask for an external VCF, and just suggest in the documentation that this can easily be done with bbtools (give command)
If a GenomicAlignments or GRanges object is passed instead of the path to a BAM file, this will just try excluding reads that aren't part of the provided regions. This is helpful because
That way InvertypeR can be given either a GAlignments composite file (to throw away useless reads) or a BAM composite file (old-school InvertypeR)
}
