% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_reads.R
\name{read_regions}
\alias{read_regions}
\title{Reads BAM files into R}
\usage{
read_regions(
  WW_bam,
  WC_bam,
  regions,
  paired_reads = TRUE,
  blacklist = blacklist
)
}
\arguments{
\item{WW_bam}{Path to the WW composite files}

\item{WC_bam}{Path to the WC composite file}

\item{regions}{A GRanges object for putative inversions}

\item{paired_reads}{A boolean: are the reads paired-end or not? Default TRUE.}

\item{blacklist}{Path to a BED file of regions to exclude}
}
\value{
A list containing two GRanges objects of reads
}
\description{
Reads near or within putative inversions (regions) that do not overlap the blacklist are read into GRanges objects.
}
\details{
Requires mapQ>=10 and excludes unmapped and duplicate reads.
}
