% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{startTimedMessage}
\alias{startTimedMessage}
\title{Timing function from BreakpointR/AneuFinder (but not exported)
Credit due to the authors of those packages.}
\usage{
startTimedMessage(...)
}
\description{
Timing function from BreakpointR/AneuFinder (but not exported)
Credit due to the authors of those packages.
}
