% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{stopTimedMessage}
\alias{stopTimedMessage}
\title{A second timing function from BreakpointR/AneuFinder (but not exported)
Credit due to the authors of those packages.}
\usage{
stopTimedMessage(ptm)
}
\description{
A second timing function from BreakpointR/AneuFinder (but not exported)
Credit due to the authors of those packages.
}
