% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twopeaks.R
\name{twopeaks}
\alias{twopeaks}
\title{Finds two peaks in a vector}
\usage{
twopeaks(x)
}
\arguments{
\item{x}{A vector of non-negative integers.}
}
\value{
A vector of length two.
}
\description{
Looks for the two biggest peaks (local maxima) in a vector of non-negative integers.
}
\details{
This is used to determine the start and end coordinates of an inversion from a list of deltaW values, which are associated with reads. It uses run length encoding
to find the two highest local maxima. It doesn't select two adjacent peaks: instead, it introduces a spacer equal to 1/12th the length of the vector. It returns the
coordinates of the right side of the leftmost peak and the left side of the rightmost peak.
}
\examples{

twopeaks(c(1, 2, 4, 2, 1, 3, 3, 3, 5, 5, 5, 6, 6, 2, 2, 2, 5, 4, 5, 4, 7))

}
