% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{widen}
\alias{widen}
\title{Widens GenomicRanges objects without falling off the start or end of a chromosome}
\usage{
widen(granges, seqlengths, distance)
}
\arguments{
\item{granges}{A GRanges object to widen}

\item{seqlengths}{A named vector of chromosome lengths, such as is produced by seqlengths(granges)}

\item{distance}{A positive integer: each interval will be widened by this distance to the right and to the left}
}
\value{
A widened GRanges object
}
\description{
Widens GenomicRanges objects without falling off the start or end of a chromosome
}
