% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WWCC_background.R
\name{WWCC_background}
\alias{WWCC_background}
\title{Estimates the background}
\usage{
WWCC_background(WW_reads, binsize = 1e+06, paired_reads = TRUE)
}
\arguments{
\item{WW_reads}{A GenomicAlignments object, the WW or CC composite file}

\item{binsize}{How big should the bins be? Default 1000000.}

\item{paired_reads}{A Boolean: are the reads paired-end? Default TRUE.}
}
\value{
A list: The background, the strand state of the composite file, and the number of reads in the composite file.
}
\description{
The background of a Strand-seq library is the proportion of non-directional reads, e.g., that did not incorporate BrdU and are randomly forward or reverse.
}
\details{
This function chops up the genome of a WW or CC composite file (we use the term WW composite file for convenience) into bins (default size 1 Mb). Then it counts the number of forward and reverse reads
to calculate the fraction of that map in the reverse orientation. It uses kernel density estimation to find the mode value of that fraction. The result
is either the background or \code{1 - background}, and since the background should be <0.1 we can both figure out which is which and decide what the strand state
of the composite file is (WW or CC). The advantage of using the mode here is that it will ignore outlier bins that contain large inversions or misaligned regions.
}
