% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjust_intervals.R
\name{adjust_intervals}
\alias{adjust_intervals}
\title{Wrapper for methods to adjust inversion breakpoints}
\usage{
adjust_intervals(
  inversions,
  reads,
  confidence = 0.95,
  base,
  haploid_chromosomes = NULL,
  prior,
  haploid_prior = c(0.5, 0.5),
  adjust_method = "deltas",
  paired_reads = TRUE
)
}
\arguments{
\item{inversions}{A dataframe of inversions output by genotype_inversions.}

\item{reads}{A list of two Granges objects: reads from a WW composite file and a WC composite file.}

\item{confidence}{Posterior probability threshold above which you consider genotype calls to be reliable. Used to decide whether to keep adjusted inversions, as well as to identify low-confidence calls
for adjust_method "low". Default 0.95.}

\item{base}{A list output by WW_background().}

\item{haploid_chromosomes}{A vector of the names of chromosomes expected to be haploid (e.g., chrX and chrY in human males). Default NULL.}

\item{prior}{Vector of three prior weights for inversion genotypes. For example, c("ref","het","hom") = c(0.9,0.05,0.05).}

\item{haploid_prior}{Vector of two prior weights for haploid chromosomes, such as chrX and chrY in human males. For example, c("ref", "inv") = c(0.9,0.1). Default c(0.5,0.5).}

\item{paired_reads}{Boolean: are the reads paired-end? Default TRUE.}

\item{adjust_method.}{See invertyper() for more details. Default "deltas".}
}
\value{
A dataframe with adjusted inversions.
}
\description{
Provides a few options as to how to adjust inversion start and end coordinates.
}
\details{
See invertyper() for more details.
}
