% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bamregion2GRanges_for_invertyper.R
\name{bamregion2GRanges_for_invertyper}
\alias{bamregion2GRanges_for_invertyper}
\title{Overrides StrandPhaseR's function for loading sequence data}
\usage{
bamregion2GRanges_for_invertyper(
  bamfile,
  bamindex,
  region = NULL,
  pairedEndReads = FALSE,
  min.mapq = 10,
  filterAltAlign = TRUE,
  all_alignments = galignmentslist_global_for_invertyper
)
}
\arguments{
\item{bamindex}{Bam-index file with or without the .bai ending. If this file does not exist it will be created and a warning is issued. (Not needed for InvertypeR)}

\item{region}{If only a subset of the genomic regions should be loaded.}

\item{pairedEndReads}{Set to \code{TRUE} if you have paired-end reads in your file.}

\item{min.mapq}{Minimum mapping quality when importing from BAM files. (not actually used by InvertypeR, same as filterAltAlign)}

\item{filterAltAlign}{(A parameter used by StrandPhaseR but accepted and then ignored by InvertypeR).}

\item{all_alignments}{A list of GAlignments objects for the Strand-seq libraries. Global so that there are no problems accessing it in parallel}

\item{file}{Bamfile with aligned reads. (Or in InvertypeR, the filename of a bamfile)}
}
\value{
a GRanges object
}
\description{
This replaces the StrandPhaseR function by sneakily preventing it from reading from the BAM file, but instead giving it processed GAlignments results
A list of GAlignments objects, named by BAM file name, will need to be available to this function even though it's not passed from StrandPhaseR
}
