% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/breakpointr_for_invertyper.R
\name{breakpointr_for_invertyper}
\alias{breakpointr_for_invertyper}
\title{A stripped down parallel implementation of the core BreakpointR function, "runBreakpointr".}
\usage{
breakpointr_for_invertyper(
  grangeslist = NULL,
  plotspath = NULL,
  numCPU = 4,
  windowsize = 1e+06,
  binMethod = "size",
  minReads = 50,
  background = 0.2,
  maskRegions = NULL,
  chromosomes = NULL
)
}
\arguments{
\item{grangeslist}{List of GRanges objects containing single-end reads (or the first read from paired reads)}

\item{plotspath}{File path where PDFs should be generated. Default NULL}

\item{numCPU}{See BreakpointR}

\item{windowsize}{See BreakpointR}

\item{binMethod}{See BreakpointR}

\item{minReads}{See BreakpointR}

\item{background}{See BreakpointR}

\item{maskRegions}{See BreakpointR}

\item{chromosomes}{See BreakpointR}
}
\value{
A Breakpoints object from BreakpointR
}
\description{
Lots of this code is borrowed/altered from the BreakpointR package by David Porubsky, Ashley Sanders, and Aaron Taudt.
It doesn't do such nice validation of arguments etc., but it is sufficient for InvertypeR's purposes
}
