% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discover_possible_inversions.R
\name{discover_possible_inversions}
\alias{discover_possible_inversions}
\title{It is currently unwise to use this function with few chromosomes, because if no reads in a library overlap the blacklist then that library will be skipped will probably cause an error}
\usage{
discover_possible_inversions(
  composite_files = list(),
  windowsize = c(40, 120, 260),
  minReads = c(15, 50, 50),
  paired_reads = FALSE,
  numCPU = 24,
  chromosomes = NULL,
  blacklist = NULL,
  background = 0.2
)
}
\arguments{
\item{background}{}
}
\value{
\link{...}
}
\description{
It is currently unwise to use this function with few chromosomes, because if no reads in a library overlap the blacklist then that library will be skipped will probably cause an error
}
