% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/galignment_to_granges.R
\name{galignment_to_granges}
\alias{galignment_to_granges}
\title{Converting reads for various tools
The 'region' argument is only for StrandPhaseR
The purpose argument mus tbe either 'StrandPhaseR' or 'BreakpointR'
This function should be inserted into InvertypeR whenever these task are done, for example, in the process_reads and pair2frgm functions
The breakpointR purpose here pairs fragments when we're looking at paired reads}
\usage{
galignment_to_granges(
  galignments = NULL,
  purpose = NULL,
  paired_reads = TRUE,
  region = NULL,
  pair2frgm = FALSE
)
}
\arguments{
\item{pair2frgm}{}
}
\value{
\link{...}
}
\description{
Converting reads for various tools
The 'region' argument is only for StrandPhaseR
The purpose argument mus tbe either 'StrandPhaseR' or 'BreakpointR'
This function should be inserted into InvertypeR whenever these task are done, for example, in the process_reads and pair2frgm functions
The breakpointR purpose here pairs fragments when we're looking at paired reads
}
