% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjustment.R
\name{genotype_new_regions}
\alias{genotype_new_regions}
\title{Checks if adjusted inversions have the same genotype}
\usage{
genotype_new_regions(
  new_regions,
  het,
  hom,
  reads,
  confidence,
  base,
  sex = "female",
  prior,
  prior_male,
  paired_reads = TRUE
)
}
\arguments{
\item{new_regions}{A list of length two: adjust_deltaW() output for the het and hom cases.}

\item{het}{A GRanges object of heterozygous inversions.}

\item{hom}{A GRanges object of homozygous inversions.}

\item{reads}{reads A list of two Granges objects: reads from a WW composite file and a WC composite file.}

\item{confidence}{Posterior probability threshold above which you consider genotype calls to be reliable. Used to decide whether to keep adjusted inversions. Default 0.95.}

\item{base}{A list output by WWCC_background().}

\item{sex}{Sex of sample to figure out sex chromosomes. Default "female".}

\item{prior}{Vector of three prior weights for inversion genotypes. For example, c("ref","het","hom") = c(0.9,0.05,0.05).}

\item{prior_male}{Vector of two prior weights for male sex chromosomes. For example, c("ref", "inv") = c(0.9,0.1).}

\item{paired_reads}{Boolean: are the reads paired-end? Default TRUE.}
}
\value{
A list containing two GRanges objects: the adjusted heterozygous inversions and the adjusted homozygous inversions.
}
\description{
Runs genotype_inversions(), discards inversions that don't have the same genotype as before adjustment, and then adds back in the originals for the ones that weren't adjusted.
}
