% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/invertyper.R
\name{invertyper}
\alias{invertyper}
\title{Genotypes inversions and adjusts their start and end coordinates.}
\usage{
invertyper(
  WW_reads,
  WC_reads,
  regions_to_genotype,
  blacklist = NULL,
  paired_reads = TRUE,
  haploid_chromosomes = NULL,
  confidence = 0.95,
  prior = c(0.333, 0.333, 0.333),
  haploid_prior = c(0.5, 0.5),
  chromosomes = NULL,
  output_file = "inversions.txt",
  adjust_method = c("raw", "merge", "deltas", "minimal", "low", "all"),
  output_folder = "."
)
}
\arguments{
\item{WW_reads}{A GenomicAlignments or GenomicAlignmentPairs object of reads from a Strand-seq WW composite file for an individual. Alternatively, the path to a BAM-formatted WW composite file.}

\item{WC_reads}{A GenomicAlignments or GenomicAlignmentPairs object of reads from a Strand-seq WC composite file for an individual. Alternatively, the path to a BAM-formatted WC composite file.}

\item{regions_to_genotype}{A GRanges object containing genomic intervals to be genotyped (putative inversions).}

\item{blacklist}{A GRanges containing regions genomic intervals with poor-quality Strand-seq data. Reads that overlap these intervals will not be used. Highly recommended. Default "".}

\item{paired_reads}{Boolean: are the reads paired-end? Default TRUE.}

\item{haploid_chromosomes}{A vector of the names of chromosomes expected to be haploid (e.g., chrX and chrY in human males). Default NULL.}

\item{confidence}{Posterior probability threshold above which you consider genotype calls to be reliable. Used to decide whether to keep adjusted inversions, as well as to identify low-confidence
calls for adjust_method "low". Default 0.95.}

\item{prior}{Vector of three prior weights for inversion genotypes. For example, c("REF","HET","HOM") = c(0.9,0.05,0.05). Default c(0.33,0.33,0.33).}

\item{haploid_prior}{Vector of two prior weights for inversions on haploid chromosomes, like chrX and chrY in human males. For example, c("REF", "INV") = c(0.9,0.1). Default c(0.5,0.5).}

\item{chromosomes}{Vector of chromosome names to restrict the search for inversions.}

\item{output_file}{Name of the file to write to. Default "inversions.txt".}

\item{adjust_method}{One of "raw", "merge", "deltas", "minimal", "low", or "all". Specifies which method to use to adjust the inversion coordinates (start- and end-points). The adjustment routine
ensures that adjusted inversions have the same genotype as they did before adjustment (if applicable), and that they overlap at least one of the original unadjusted inversions in every cluster of
overlapping events. If after adjustment (except with "raw") overlapping inversions still remain, we merge the ones of the same genotype. If there are still overlapping inversions, we take the largest.
"raw": No adjustment. "merge": Merge overlapping confident (i.e., posterior probability >= confidence for a genotype) inversions of the same genotype. "Deltas": adjust the endpoints of confident
inversions based on the Strand-seq data. We calculate deltaW values for each read, which measure the change in read direction for nearby reads, and take the coordinates that correspond to the two
highest values (i.e. the two spots with the greates change in read direction, which we hope will correspond to inversion breakpoints). This is done once for every set of overlapping inversions of the
same genotype. "minimal": An experimental method. For confident overlapping inversions of the same genotype, take the interval common to all of them if one exists. "low": Much like "deltas", except
that we only adjust inversions for which no confident genotype could be found (i.e., all prior probabilities < confidence). Sometimes this will "find" a confident inversion if the original coordinates
were slightly off. "all": Does both "deltas" and "low".}
}
\value{
A dataframe that associates each entry regions_to_genotype (putative inversions) with a genotype and posterior probability. Cols 1-3 are genomic coordinates, and cols 4-7 are read counts for the
two composite files (WC and WW, or WC and CC if that's what you submitted) subset by Watson (W) and Crick (C) strands. Col 8 is the most probable genotype, and col 9 is the posterior probability
associated with that genotype. Col 10 labels inversions with low read density. The start and end coordinates of such inversions should be checked by viewing Strand-seq data in a genome browser.
}
\description{
Requires WW and WC composite files (from create_composite_files() or in BAM format) and a GRanges object containing putative inversion coordinates
}
\details{
We highly recommend using a blacklist, and we provide one on the GitHub page for humans. For non-humans, regions that have anomalously high depth of coverage in many cells or that always have WC
(Watson-Crick) strand-state are suspect and should be added to a blacklist by some custom analysis (for example, see the sup mat of https://doi.org/10.1186/s12864-021-07892-9).

Priors should chosen to reflect the number of inversions you expect to find in an individual (up to ~100 if you use, for example, all publised inversion coordinates (2020)) and the number of unique
your putative inversions and you're just starting out, use adjust_method="raw". Otherwise, use adjust_method="deltas" or "all". The other options are specific to particular use cases.

Note that we can phase inversions only because the WC composite file already has phased reads. This means that we know a 0|1 inversion on chr1 is on the same homolog as all 0|1 inversions
on chr1 in the sample, and that all chr1 1|0 inversions are on the other homolog. However, we don't know whether a 0|1 inversion on chr1 and a 0|1 inversion chr2 came from the same parent.
0|1 inversions are distinguished from 1|0 inversions based on the strand switch in the WC composite file ( WC -> WW or WC -> CC).
}
