% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phaseChromosome_for_invertyper.R
\name{phaseChromosome_for_invertyper}
\alias{phaseChromosome_for_invertyper}
\title{A stripped-down version of the StrandPhaseR function phaseChromosome by David Porubsky}
\usage{
phaseChromosome_for_invertyper(
  input_folder,
  output_folder = "./StrandPhaseR_analysis",
  positions = NULL,
  WCregions = NULL,
  chromosome = NULL,
  pairedEndReads = TRUE,
  min.mapq = 10,
  min.baseq = 30,
  num.iterations = 2,
  translateBases = TRUE,
  fillMissAllele = NULL,
  splitPhasedReads = FALSE,
  compareSingleCells = FALSE,
  exportVCF = NULL,
  galignmentslist = NULL
)
}
\arguments{
\item{input_folder}{Path to the bam files to process}

\item{output_folder}{Output directory.}

\item{positions}{Filename with listed position of SNVs for given chromosome (format: chrName SNVpos).}

\item{WCregions}{Filename of all WC region for a given chromosome (format: chrName:Start:End:FileName).}

\item{chromosome}{If only a subset of the chromosomes should be processed, specify them here.}

\item{pairedEndReads}{Set to \code{TRUE} if you have paired-end reads in your file.}

\item{min.mapq}{Minimum mapping quality when importing from BAM files.}

\item{min.baseq}{Minimum base quality to consider a base for phasing.}

\item{num.iterations}{Number of iteration to sort watson and crick matrices.}

\item{translateBases}{...}

\item{splitPhasedReads}{Set to \code{TRUE} if you want to split reads per haplotype.}

\item{compareSingleCells}{Set to \code{TRUE} if you want to compare haplotypes at the single-cell level.}

\item{exportVCF}{...}

\item{fillGaps}{...}
}
\value{
A subset of the StrandPhaseR output that phases the strand state of wc regions
}
\description{
This is a simplified function that removes some of the functionality (and runtime) of David's original StrandPhaseR function
The arguments are the same (or effectively the same) as David's corresponding function.
}
\details{
The descriptions of arguments here are just copied from StrandPhaseR
}
