% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_UCSC_browser_files.R
\name{write_UCSC_browser_files}
\alias{write_UCSC_browser_files}
\title{Writes genome browser files for inversions}
\usage{
write_UCSC_browser_files(
  inversions = NULL,
  WW_reads = NULL,
  WC_reads = NULL,
  confidence = 0.95,
  paired_reads = TRUE,
  outputfolder = "./",
  prefix = ""
)
}
\arguments{
\item{prefix}{}
}
\value{
\link{...}
}
\description{
Given a list of inversions from invertyper() and two composite files, it makes nicely-formatted BED files compatible with the UCSC Genome Browser
The inversions are coloured by genotype (black is 1/1, grey is 0/1, light grey is 0/0), and the Strand-seq reads are coloured by strand as in BreakpointR etc.
}
